self.x = torch.from_numpy(x).float()         self.y = torch.from_numpy(y).float()

+++ ptlk/data/datasets.py
@@ -29,8 +29,8 @@ class ModelNet(globset.Globset):
 
 class Voltages(Dataset):
     def __init__(self, x, y):
-        self.x = torch.from_numpy(x)
-        self.y = torch.from_numpy(y)
+        self.x = torch.from_numpy(x).float()
+        self.y = torch.from_numpy(y).float()
 
     def __getitem__(self, idx):
         return self.x[idx], self.y[idx]
